Updating for better Ubuntu/Debian installation
This is in response to update-rc.d not working, period, without the runlevels in the init.d script and the "enabled" option on the update-rc.d line. The update-rc.d man page specifically says that "enable" only:

       modifies existing  runlevel  links  for  the
       script  /etc/init.d/name  by  renaming stop links to start
       links with a sequence number equal to the positive differ‐
       ence  of current sequence number minus 100, thus returning
       to the original sequence number that the script  had  been
       installed with before disabling it.

I am unsure of what exactly the proper "Default-start" and "Default-Stop" run levels need to be, or if any other items should be added to the "Required-Start" and "-Stop" line items in the init.d script.

Without the correct runtime items, the script stops at that point, and doesn't continue with the "userify-start" script.