Cosmetic Adjustments, Suggestions by Comment
- In the directions at the top, added `-sE`, as this is whats listed at
https://userify.com/docs/enterprise/installation-enterprise/ to use.
- Removed the echo’s from the $URL block and replaced with EOF. Also
adjusted for readability. Added `-r` flag to `read` just to make sure
nothing tries to `newline` itself in there.
- removed the space from `#!` and `/bin/bash` at the
userify-server-init, as shebangs are supposed to be `#!/bin/`. This is
just to be SURE nothing gets hinky in the future.
- added `chmod +x` under the init-script, since we’re adding to
autostart here. Also added possible if/elif conditionals by comments.
- changed the EOF from ` to “ just for continuity’s sake
- Added quotes to the bottom echo’s
- removed the & from the last command, as this would stop the install
script, but still run the server script in the background with the
output in the STDOUT output without being able to `ctrl-c` the server.